We looked at this together on Friday. Good to merge!

+++ src/http_headers/HttpHeaders.cpp
@@ -3,7 +3,6 @@
 #include <cstddef>
 #include <cstring>
 #include <string>
-#include <sys/types.h>
 
 static const struct {
   std::string extension;
@@ -53,12 +52,12 @@ void HttpHeaders::unsetAll(void) {
   _headersSet = 0;
 }
 
-void HttpHeaders::setContentLength(const off_t Length) {
+void HttpHeaders::setContentLength(const size_t Length) {
   _headersSet |= ContentLength;
   _contentLength = Length;
 }
 
-off_t HttpHeaders::getContentLength(void) const {
+size_t HttpHeaders::getContentLength(void) const {
   return _contentLength;
 }
